Move MainActivity into activities package
