+++ app/src/main/AndroidManifest.xml
@@ -16,7 +16,7 @@
         tools:targetApi="31"
         android:networkSecurityConfig="@xml/network_security_config">
         <activity
-            android:name=".MainActivity"
+            android:name=".activities.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
